+++ src/migration/scripts/8-add-rank-to-pvpstandings.js
@@ -1,0 +1,8 @@
+const newFields = require('../lib/newFields');
+
+const fields = (Sequelize = {}) => ({
+  naRank: Sequelize.INTEGER,
+  gw2aRank: Sequelize.INTEGER,
+});
+
+module.exports = newFields(fields, 'PvpStandings');