+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## [1.3.0] - 2017/01/09
 ### Added
+- `gw2aRank` and `naRank` to `PvpStandings` table
 - Pvp leaderboard fetcher to calculate gw2aRank and save it to the database
 
 ## [1.2.2] - 2017/01/08